+++ website/src/stories/main-editor.tsx
@@ -14,6 +14,7 @@ import {
   handleLinkClick,
   handleTagClick,
   handleWikilinkClick,
+  resolveLinkPreview,
   searchNotes,
   searchTags,
 } from '../lib/demo-data.ts'
@@ -221,6 +222,7 @@ function MainEditorDemo() {
             onWikilinkSearch={searchNotes}
             onImageClick={handleImageClick}
             onLinkClick={handleLinkClick}
+            resolveLinkPreview={resolveLinkPreview}
             onTagClick={handleTagClick}
             onWikilinkClick={handleWikilinkClick}
           >